Added autorefresh feature to conversation views
[autorefresh]

This is not true dynamic HTML, but the effect is similar: if a user
has a conversation displayed (and JS enabled), their window will
refresh whenever it is necessary.

This is implemented in the traditional long poll (Comet) style.
Therefore, it will not work with a single-threaded server, eg. the
default Django development server; you need a multithreaded
installation. See for example:

http://twigstechtips.blogspot.com/2011/05/django-run-development-server-in-multi.html

Once you have taken care of that, the thing will work, but do not
expect a high performance. Traditional web servers are not designed
for high Comet loads, and they will quickly become sluggish and lock
up -- a more mature solution would require an ansynchronous server
like Orbited or Tornado.

Refreshing on incoming mails was a particularly tricky part, since
as of now, SMTP is handled by a separate process, so there was no
easy way to notify the web server except for a dedicated HTTP request.
So, right now, the SMTP process does a GET request to notify the HTTP
process about incoming mails. Of course I do not consider this
anything else but a stopgap measure.

All in all, I did this mainly to inspire myself and perhaps to be able
to show off some of the upcoming features of Hk. Hk is not
production-ready, and this feature is even less so.