Updated code to work with Django 1.5
[django-1.5]

Out of the several incompatibilities, this one was fixed:

"""One deprecated feature worth noting is the shift to “new-style” url
tag. Prior to Django 1.3, syntax like {% url myview %} was interpreted
incorrectly (Django considered "myview" to be a literal name of a
view, not a template variable named myview). Django 1.3 and above
introduced the {% load url from future %} syntax to bring in the
corrected behavior where myview was seen as a variable."""

Admin works fine, even though there were problems on 1.4 with serving static
files. Maybe they reverted this change in 1.5? Further testing on 1.4 is
necessary.